OWASP ZAP scan via Docker Compose
This has the most basic configuration for running the OWASP ZAP scanner
against a local Reaction installation using Docker Compose. This is
placed in the .circleci directory as it's intended that this would run
as part of continuous integration. README included with instructions for
running but no CI config yet.